+++ code/src/clt_mlm.h
@@ -19,7 +19,7 @@ struct clt_mlm_state {
 
 
 int
-clt_mlm_setup(struct clt_mlm_state *s, const char *dir, const long *pows,
+clt_mlm_setup(struct clt_mlm_state *s, const char *dir, const int *pows,
               long kappa, long size, int verbose);
 
 void
